refactor(argocd): manage argocd in kustomize instead of ansible
Ansible will execute the kustomize of the argocd kustomize folder if ArgoCD isn't deployed(checked by the presence of deploy/argocd-server in the argocd namespace.

This allows managing ArgoCD with ArgoCD and allows for better customization(with kustomize) and avoids deploying the default manifest, then overriding it back to what I want.